+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Updated DESCRIPTION to require purrr >= 0.1.1
 * Fixed cast_sparse, cast_dtm, and other sparse casters to ignore groups in the input (#19)
+* Changed unnest_tokens so that it no longer uses tidyr's unnest, but rather a custom version that removes some overhead. In some experiments, this sped up unnest_tokens on large inputs by about 40%. This also moves tidyr from Imports to Suggests for now.
+* unnest_tokens now checks that there are no list columns in the input, and raises an error if present (since those cannot be unnested)
 
 # tidytext 0.1.1